Enable object deserialization test in component testsuite.
This PR modifies the component test-suite to have a canKubeDeserialize
field that checks whether objects can be deserialized using the internal
Kubernetes deserialization logic. This relies on the schema in the
patchtmpl (which perhaps should be put in a more common place, like
`pkg/options`)

A couple smaller changes:

- disable jsonnet importing example for now
- add jsonnet to the multi-applier
- add a deserialize test case to all examples
- add a test case for jsonnet
