+++ examples/patchbuilder/test-suite.yaml
@@ -1,6 +1,16 @@
 componentFile: builder.yaml
-rootDirectory: './'
+
 testCases:
+- description: Deserialize Objects
+  build:
+    options:
+      ImageTag: '2.0'
+  apply:
+    options:
+      Namespace: default
+  expect:
+    canKubeDeserialize: true
+
 - description: Success
   build:
     options:
