+++ pkg/options/multi/multi.go
@@ -20,6 +20,7 @@ import (
 	bundle "github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/apis/bundle/v1alpha1"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/options"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/options/gotmpl"
+	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/options/jsonnet"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/options/patchtmpl"
 )
 
@@ -37,6 +38,7 @@ func NewDefaultApplier() options.Applier {
 	return NewApplier([]options.Applier{
 		gotmpl.NewApplier(),
 		patchtmpl.NewDefaultApplier(),
+		jsonnet.NewApplier(nil /* importer */),
 	})
 }
