+++ pkg/testutil/componentsuite/BUILD.bazel
@@ -5,6 +5,7 @@ go_library(
     srcs = [
         "componentsuite.go",
         "schema.go",
+        "validate.go",
     ],
     importpath = "github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/testutil/componentsuite",
     visibility = ["//visibility:public"],
@@ -14,10 +15,13 @@ go_library(
         "//pkg/converter:go_default_library",
         "//pkg/filter:go_default_library",
         "//pkg/options:go_default_library",
+        "//pkg/options/multi:go_default_library",
         "//pkg/options/patchtmpl:go_default_library",
         "//pkg/testutil:go_default_library",
         "//pkg/validate:go_default_library",
         "//pkg/wrapper:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/apis/meta/v1/unstructured:go_default_library",
+        "//vendor/k8s.io/apimachinery/pkg/runtime:go_default_library",
     ],
 )
