+++ pkg/testutil/componentsuite/componentsuite.go
@@ -17,20 +17,17 @@ package componentsuite
 
 import (
 	"context"
-	"fmt"
 	"io/ioutil"
 	"os"
 	"path/filepath"
-	"strings"
 	"testing"
 
 	bundle "github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/apis/bundle/v1alpha1"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/build"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/converter"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/filter"
-	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/options/patchtmpl"
+	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/options/multi"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/testutil"
-	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/validate"
 	"github.com/GoogleCloudPlatform/k8s-cluster-bundle/pkg/wrapper"
 )
 
@@ -130,88 +127,20 @@ func runApply(t *testing.T, comp *bundle.Component, tc *TestCase) *bundle.Compon
 	applyOpts := &filter.Options{}
 	applyOpts.Annotations = tc.Apply.Filters
 
-	applier := patchtmpl.NewApplier(
-		patchtmpl.DefaultPatcherScheme(),
-		applyOpts,
-		true /* includeTemplates, for debugging */)
+	applier := multi.NewDefaultApplier()
 
 	comp, err := applier.ApplyOptions(comp, tc.Apply.Options)
 	cerr := testutil.CheckErrorCases(err, tc.Expect.ApplyErrSubstr)
 	if cerr != nil {
 		t.Fatal(cerr)
 	}
+
+	o, _ := converter.FromObject(comp).ToYAMLString()
+	t.Logf("obj: %v", o)
+
 	if err != nil {
 		// Since there's an error, it's a terminal condition
 		return nil
 	}
 	return comp
-}
-
-type objKey struct {
-	name string
-	kind string
-}
-
-func (ok objKey) String() string {
-	return fmt.Sprintf("{name: %q, kind: %q}", ok.name, ok.kind)
-}
-
-func runValidate(t *testing.T, comp *bundle.Component, tc *TestCase) {
-	if errList := validate.Component(comp); len(errList) > 0 {
-		t.Errorf("There were errors validating component:\n%v", errList.ToAggregate())
-	}
-
-	objCheckMap := make(map[objKey]ObjectCheck)
-	for _, oc := range tc.Expect.Objects {
-		objCheckMap[objKey{
-			name: oc.Name,
-			kind: oc.Kind,
-		}] = oc
-	}
-
-	objMap := make(map[objKey]string)
-	for _, obj := range comp.Spec.Objects {
-		matchFail := false
-
-		key := objKey{name: obj.GetName(), kind: obj.GetKind()}
-		objStr, err := converter.FromObject(obj).ToYAMLString()
-		if err != nil {
-			// This is a very unlikely error.
-			t.Fatal(err)
-		}
-		objMap[key] = objStr
-
-		check := objCheckMap[key]
-		for _, expStr := range check.FindSubstrs {
-			if !strings.Contains(objStr, expStr) {
-				t.Errorf("Did not find %q in object %v, but expected to", expStr, key)
-				matchFail = true
-			}
-		}
-
-		for _, noExpStr := range check.NotFindSubstrs {
-			if strings.Contains(objStr, noExpStr) {
-				t.Errorf("Found %q in object %v, but did not expect to", noExpStr, key)
-				matchFail = true
-			}
-		}
-
-		if matchFail {
-			t.Errorf("Contents for object that didn't meet expectations %v:\n%s", key, objStr)
-		}
-	}
-
-	for key := range objCheckMap {
-		if _, ok := objMap[key]; !ok {
-			t.Errorf("Got object-keys %q, but expected to find object %v", stringMapKeys(objMap), key)
-		}
-	}
-}
-
-func stringMapKeys(m map[objKey]string) []string {
-	var out []string
-	for k := range m {
-		out = append(out, k.String())
-	}
-	return out
 }
